+++ CodeWars/Algorithms/SumByFactors/Java/src/SumOfDividedTest.java
@@ -1,0 +1,13 @@
+
+import static org.junit.Assert.*;
+import org.junit.*;
+
+public class SumOfDividedTest {
+
+    @Test
+    public void testOne() {
+        int[] lst = new int[] {12, 15};
+        assertEquals("(2 12)(3 27)(5 15)",
+                SumOfDivided.sumOfDivided(lst));
+    }
+}